requirements.txt
smolagents>=0.1.0
azure-cognitiveservices-vision-computervision>=0.9.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation>=0.42b0
python-dotenv>=1.0.0
pytest>=7.4.0
rich>=13.7.0
pydantic>=2.5.0 